edit penambahan role di user

+++ back-end/src/entities/pokja/routes.js
@@ -1,5 +1,6 @@
 const router = require("express").Router();
 const { tambahPokja, listPokja, hapusPokja } = require("./controller");
+const checkAdmin = require("../../modules/checkAdmin");
 
 router
   .route("/")
@@ -12,30 +13,38 @@ router
     );
   })
   .post((rq, rs) => {
-    tambahPokja(rq.body.pokja)
-      .then(() =>
-        rs.status(201).json({
-          message: "Pokja ditambah",
-          data: rq.body
-        })
-      )
-      .catch(err => {
-        rs.status(500).json({ message: "Input salah atau pokja sudah ada" });
-      });
+    if (checkAdmin(rq.signedCookies.usertoken)) {
+      tambahPokja(rq.body.pokja)
+        .then(() =>
+          rs.status(201).json({
+            message: "Pokja ditambah",
+            data: rq.body
+          })
+        )
+        .catch(err => {
+          rs.status(500).json({ message: "Input salah atau pokja sudah ada" });
+        });
+    } else {
+      rs.status(401);
+    }
   });
 
 router.route("/:id").delete((rq, rs) => {
-  hapusPokja(rq.params.id).then(val => {
-    if (val === 1) {
-      rs.status(200).json({
-        message: "Pokja dihapus",
-        where: rq.params
-      });
-    } else {
-      rs.status(404).json({
-        message: "ID pokja tidak ditemukan"
-      });
-    }
-  });
+  if (checkAdmin(rq.signedCookies.usertoken)) {
+    hapusPokja(rq.params.id).then(val => {
+      if (val === 1) {
+        rs.status(200).json({
+          message: "Pokja dihapus",
+          where: rq.params
+        });
+      } else {
+        rs.status(404).json({
+          message: "ID pokja tidak ditemukan"
+        });
+      }
+    });
+  } else {
+    rs.status(401);
+  }
 });
 module.exports = router;
